Refactor view model
